move files under domain folder
