+++ lib/domain/ingredients/ingredients_stock.dart
@@ -1,5 +1,5 @@
-import 'package:pizza_coop/ingredients/ingredient.dart';
-import 'package:pizza_coop/ingredients/ingredients_catalog.dart';
+import 'package:pizza_coop/domain/ingredients/ingredient.dart';
+import 'package:pizza_coop/domain/ingredients/ingredients_catalog.dart';
 
 class IngredientsStock {
   final List<StockIngredient> _ingredients = [];
