+++ test/domain/ingredients/ingredients_stock_test.dart
@@ -1,6 +1,6 @@
-import 'package:pizza_coop/ingredients/ingredient.dart';
-import 'package:pizza_coop/ingredients/ingredients_catalog.dart';
-import 'package:pizza_coop/ingredients/ingredients_stock.dart';
+import 'package:pizza_coop/domain/ingredients/ingredient.dart';
+import 'package:pizza_coop/domain/ingredients/ingredients_catalog.dart';
+import 'package:pizza_coop/domain/ingredients/ingredients_stock.dart';
 import 'package:test/test.dart';
 
 void main() {
