+++ test/domain/stock_role_test.dart
@@ -1,8 +1,8 @@
-import 'package:pizza_coop/ingredients/ingredient.dart';
-import 'package:pizza_coop/ingredients/ingredients_catalog.dart';
-import 'package:pizza_coop/ingredients/ingredients_stock.dart';
-import 'package:pizza_coop/stock_role.dart';
-import 'package:pizza_coop/wallet.dart';
+import 'package:pizza_coop/domain/ingredients/ingredient.dart';
+import 'package:pizza_coop/domain/ingredients/ingredients_catalog.dart';
+import 'package:pizza_coop/domain/ingredients/ingredients_stock.dart';
+import 'package:pizza_coop/domain/stock_role.dart';
+import 'package:pizza_coop/domain/wallet.dart';
 import 'package:test/test.dart';
 
 void main() {
