+++ test/domain/wallet_test.dart
@@ -1,4 +1,4 @@
-import 'package:pizza_coop/wallet.dart';
+import 'package:pizza_coop/domain/wallet.dart';
 import 'package:test/test.dart';
 
 void main() {
